+++ tests/objects/test_matrix44.py
@@ -1,9 +1,13 @@
 from __future__ import absolute_import
+
 try:
     import unittest2 as unittest
 except:
     import unittest
 import numpy as np
+
+from numpy import float32
+
 from pyrr.objects.matrix33 import Matrix33
 from pyrr.objects.matrix44 import Matrix44
 from pyrr.objects.quaternion import Quaternion
@@ -292,34 +296,34 @@ class test_object_matrix44(unittest.TestCase):
         # define expectations for multiple cases
         testsets = [
             (
-                Vector3([1, 1, 2], dtype='f4'),
-                Quaternion.from_y_rotation(np.pi, dtype='f4'),
-                Vector3([10, 0, -5], dtype='f4'),
+                Vector3([1, 1, 2], dtype=float32),
+                Quaternion.from_y_rotation(np.pi, dtype=float32),
+                Vector3([10, 0, -5], dtype=float32),
                 Matrix44([
                     [-1, 0, 0, 0],
                     [0, 1, 0, 0],
                     [0, 0, -2, 0],
                     [10, 0, -5, 1],
-                ], dtype='f4')
+                ], dtype=float32)
             ),
             (
-                Vector3([-1, 3, .5], dtype='f4'),
-                Quaternion.from_axis_rotation(Vector3([.75, .75, 0], dtype='f4').normalized, np.pi, dtype='f4').normalized,
-                Vector3([1, -1, 1], dtype='f4'),
+                Vector3([-1, 3, .5], dtype=float32),
+                Quaternion.from_axis_rotation(Vector3([.75, .75, 0], dtype=float32).normalized, np.pi, dtype=float32).normalized,
+                Vector3([1, -1, 1], dtype=float32),
                 Matrix44([
                     [0, -1, 0, 0],
                     [3, 0, 0, 0],
                     [0, 0, -.5, 0],
                     [1, -1, 1, 1],
-                ], dtype='f4')
+                ], dtype=float32)
             ),
         ]
 
         for expected_scale, expected_rotation, expected_translation, expected_model in testsets:
             # compose model matrix using original inputs
-            s = Matrix44.from_scale(expected_scale, dtype='f4')
-            r = Matrix44.from_quaternion(expected_rotation, dtype='f4')
-            t = Matrix44.from_translation(expected_translation, dtype='f4')
+            s = Matrix44.from_scale(expected_scale, dtype=float32)
+            r = Matrix44.from_quaternion(expected_rotation, dtype=float32)
+            t = Matrix44.from_translation(expected_translation, dtype=float32)
             m = t * r * s
 
             # check that it's the same as the expected matrix
@@ -329,9 +333,9 @@ class test_object_matrix44(unittest.TestCase):
 
             # decompose this matrix and recompose the model matrix from the decomposition
             ds, dr, dt = m.decompose()
-            ds = Matrix44.from_scale(ds, dtype='f4')
-            dr = Matrix44.from_quaternion(dr, dtype='f4')
-            dt = Matrix44.from_translation(dt, dtype='f4')
+            ds = Matrix44.from_scale(ds, dtype=float32)
+            dr = Matrix44.from_quaternion(dr, dtype=float32)
+            dt = Matrix44.from_translation(dt, dtype=float32)
             dm = dt * dr * ds
 
             # check that it's the same as the original matrix
